+++ FavoritePlaces/src/edu/umn/kill9/places/model/EventDataSource.java
@@ -16,7 +16,8 @@ import java.util.List;
  * To change this template use File | Settings | File Templates.
  */
 public class EventDataSource {
-    public Event createEvent(Place place, int calendarId) throws ParseException {
+
+    public Event createEvent(Place place, long calendarId) throws ParseException {
         ParseObject parseObject = new ParseObject(Event.TABLE_EVENT);
         parseObject.put(Event.COLUMN_CALENDARID, calendarId);
